Feature: automatic prover configuration
Problem: the user has to provide the prover configuration and parameters
when running `execute_and_prove`. The parameter file in particular is
not trivial to generate.

Solution: generate the config and parameters automatically (if not
specified by the caller). Now users can just execute and prove a program
by specifying the compiled program and let the server generate the
prover config and parameters. The user can still override these
parameters by specifying them directly.