Internal: script to generate a full test case (#16)
Problem: we wish to create different test cases for unit and integration
tests. Doing this manually is not trivial.

Solution: introduce a Python script that generates all test case files
based on the Cairo program file.

The new `generate_test_case.py` script takes a Cairo v0 program as
input, compiles it, runs it and proves it. It also generates the prover
config and parameters based on the number of steps of the program.