Feature: Starknet prover client/server
Problem: users want to execute programs/PIEs on top of the Starknet
bootloader to generate proofs compatible with Starknet.

Solution: introduce a new RPC service.

The new `StarknetProver` service enables users to execute any
combination of programs/PIEs on top of the bootloader. The proof is then
generated as usual with Stone.

This commit modifies the directory layout of several crates to better
map the use of several services.

This implementation has the following limitations:
* The bootloader is included directly in the binary, making it
  impossible to use different versions of the bootloader.
* No support for proof splitting.